requirements.txt
pycrypto==2.6.1
Django==2.0.7
ethereum==2.3.2
djangorestframework==3.8.2
validate_email==1.3
web3==4.4.1
django-cors-headers==2.3.0
django-debug-toolbar==1.9.1
django-extensions==2.1.0
django-rest-swagger==2.2.0
djangorestframework-jwt==1.11.0
